refactor: block areas clean
